+++ CHANGELOG.md
@@ -71,6 +71,10 @@ Unreleased
 - Add `MoveToOptions::visualize_path_style`to allow for path visualization of movement system.
 - Change `HasStore::store_free_capacity` to return `i32`, handling potential negative values due
   to expiration of `OPERATE_STORAGE`
+- Add `game::gcl::total_for_level` and `game::gpl::total_for_level` which calculate the total
+  lifetime points required for a given level of GCL or GPL
+- Change `constants::GCL_POW` to f64 from f32 due to slightly incorrect calculations when using
+  this from f32 to calculate GCL levels (breaking)
 
 0.7.0 (2019-10-19)
 ==================